move helper classes to outside of prototype module
